mlr3verse: Easily Install and Load the 'mlr3' Package Family

The 'mlr3' package family is a set of packages for machine-learning purposes built in a modular fashion. This wrapper package is aimed to simplify the installation and loading of the core 'mlr3' packages. Get more information about the 'mlr3' project at https://mlr3book.mlr-org.com/.

Learn mlr3

Book on mlr3: https://mlr3book.mlr-org.com

Use cases and examples gallery: https://mlr3gallery.mlr-org.com

Cheat Sheets: https://github.com/mlr-org/mlr3cheatsheets

Additional Extension Packages

The mlr3verse only installs the subset of packages which are on the one hand frequently required to perform a data analysis, and on the other hand do not pull in too many dependencies or system requirements.

Additional extension packages include:

Analysis of benchmark experiments: mlr3benchmark

Task interface for data bases: mlr3db

Connector to OpenML: mlr3oml

Additional learners: https://github.com/mlr-org/mlr3extralearners

Hyperband tuning algorithm: mlr3hyperband

Spatio-temporal resampling: mlr3spatiotempcv

See also

Author

Maintainer: Marc Becker marcbecker@posteo.de (ORCID)

Authors:

Michel Lang michellang@gmail.com (ORCID)

Patrick Schratz patrick.schratz@gmail.com (ORCID)