Print mlr3* package versions

Returns the package versions of imported and suggested mlr3 packages as a data.table::data.table(). Suggested packages that are not installed are listed with version NA.

Usage

mlr3verse_info()

Examples

mlr3verse_info()
#> Key: <package>
#>              package version
#>               <char>  <char>
#>  1:            bbotk   1.0.1
#>  2:      miesmuschel 0.0.4.2
#>  3:             mlr3  0.20.2
#>  4:    mlr3batchmark   0.1.1
#>  5:      mlr3cluster   0.1.9
#>  6:         mlr3data   0.7.0
#>  7:           mlr3db   0.5.2
#>  8:     mlr3fairness   0.3.2
#>  9:          mlr3fda   0.2.0
#> 10:      mlr3filters   0.8.0
#> 11:      mlr3fselect   1.0.0
#> 12:    mlr3hyperband   0.6.0
#> 13:     mlr3learners   0.7.0
#> 14:          mlr3mbo   0.2.4
#> 15:         mlr3misc  0.15.1
#> 16:          mlr3oml  0.10.0
#> 17:    mlr3pipelines   0.6.0
#> 18:      mlr3spatial   0.5.0
#> 19: mlr3spatiotempcv   2.3.1
#> 20:      mlr3summary   0.1.0
#> 21:        mlr3torch   0.1.0
#> 22:       mlr3tuning   1.0.0
#> 23: mlr3tuningspaces   0.5.1
#> 24:          mlr3viz   0.9.0
#> 25:          paradox   1.0.1
#> 26:             rush   0.1.1
#>              package version